Moved my test from the file "registrationUser_ValidData.spec" to already exist similar file "registrationCreateAccount" file and deleted my file
