+++ tests/registrationCreateAccount.spec.js
@@ -1,10 +1,25 @@
 import { test, expect } from "@playwright/test";
 
 test.describe("create account", () => {
-    const createAccount_url = 'https://magento.softwaretestingboard.com/customer/account/create/'
+    const createAccount_url = 'https://magento.softwaretestingboard.com/customer/account/create/';
+    let  firstName, lastName, password, passwordConfirm, email;
 
     test.beforeEach(async ({ page }) => {
       await page.goto("/");
+
+        firstName = 'Svetlana';
+        lastName = 'Kudryvzeva';
+        password = '12345Sveta!';
+        passwordConfirm = '12345Sveta!';
+
+        function generateRandomEmail() {
+            const mailbox = Math.random().toString(36).substring(2, 10);
+            const domain = "gmail.com";
+            return `${mailbox}@${domain}`;
+          }
+
+        email = generateRandomEmail();
+
     })
 
     test('should be all fields are required', async ({ page }) => {
@@ -14,7 +29,6 @@ test.describe("create account", () => {
 
     await page.getByRole('button', {name: 'Create an Account'}).click();
     const warninMessage = await page.$$('div[generated="true"]');
-
 
 
     await expect(warninMessage).toBeTruthy();
@@ -44,5 +58,28 @@ test.describe("create account", () => {
      await page.getByText('Minimum length of this field must be equal or greater than 8 symbols. Leading and trailing spaces will be ignored.').isVisible();
 
     })
+
+    test('Check that user can possible to create account with entered all valid fields', async ({ page }) => {
+      await page.getByRole('link', {name: 'Create an Account'}).click();
+      await page.locator('#firstname').click();
+      await page.locator('#firstname').fill(firstName);
+      await page.locator('#lastname').click();
+      await page.locator('#lastname').fill(lastName);
+      await page.locator('#email_address').click();
+      await page.locator('#email_address').fill(email);
+      await page.locator('#password').click();
+      await page.locator('#password').fill(password);
+      await page.locator('#password-confirmation').click();
+      await page.locator('#password-confirmation').fill(passwordConfirm);
+      await page.getByRole('button', { name: "Create an Account" }).click();
+
+
+      await expect(page.locator('.base[data-ui-id="page-title-wrapper"]')).toHaveText('My Account');
+
+      const thanksMessage = page.getByRole("alert").getByText("Thank you for registering with Main Website Store.");
+      await expect(thanksMessage).toHaveText("Thank you for registering with Main Website Store.");
+
+  })
+
 })
